<FEAT> move sms domain to domain-common module
