+++ domains/domain-common/src/main/java/com/homfo/sms/command/ValidateSmsCodeCommand.java
@@ -1,0 +1,11 @@
+package com.homfo.sms.command;
+
+/**
+ * 전화번호 인증 코드를 확인하기 위한 Command 입니다.
+ * */
+public record ValidateSmsCodeCommand(
+        String phoneNumber,
+
+        String code
+) {
+}
