+++ domains/domain-common/src/main/java/com/homfo/sms/dto/SmsSendDto.java
@@ -1,4 +1,4 @@
-package com.homfo.user.dto;
+package com.homfo.sms.dto;
 
 public record SmsSendDto(
         String phoneNumber,
