+++ domains/domain-common/src/main/java/com/homfo/sms/usecase/ValidateSmsCodeUsecase.java
@@ -1,15 +1,17 @@
-package com.homfo.user.usecase;
+package com.homfo.sms.usecase;
+
+
+import com.homfo.sms.command.ValidateSmsCodeCommand;
 
 /**
  * 사용자 전화번호가 중복되는지, 올바른 전화번호인지 확인합니다.
  * */
-public interface RequestSmsCodeUsecase {
+public interface ValidateSmsCodeUsecase {
     /**
      * 사용자 전화번호가 중복되는지, 올바른 전화번호인지 확인합니다.
      *
-     * @throws com.homfo.error.ResourceAlreadyExistException 이미 존재하는 전화번호라면
      * @throws com.homfo.error.ResourceNotFoundException 존재하지 않는 전화번호라면
      * @throws com.homfo.error.ThirdPartyUnavailableException 외부 인증 서비스에서 장애가 발생한 경우
      * */
-    boolean requestSmsCode(String phoneNumber);
+    boolean validateSmsCode(ValidateSmsCodeCommand command);
 }
